requirements.txt
dedalus-labs
dedalus-mcp
python-dotenv
httpx
